dora_digua_long/node-hub/dora-openlong-udp/pyproject.toml
[project]
name = "dora-openlong-udp"
version = "0.3.10"
authors = [{ name = "Your Name", email = "email@email.com" }]
description = "dora-openlong-udp"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "dora-rs >= 0.3.9",

    "scipy",
] # Make reachy2 sdk version configurable

[dependency-groups]
dev = ["pytest >=8.1.1", "ruff >=0.9.1"]

[project.scripts]
dora-openlong-udp = "dora_openlong.udp:main"

[tool.ruff.lint]
extend-select = [
  "D",    # pydocstyle
  "UP",   # Ruff's UP rule
  "PERF", # Ruff's PERF rule
  "RET",  # Ruff's RET rule
  "RSE",  # Ruff's RSE rule
  "NPY",  # Ruff's NPY rule
  "N",    # Ruff's N rule
  "I",    # Ruff's I rule
]
